+++ scripts/run_infrastructure.py
@@ -10,11 +10,11 @@ from typing import List
 
 DOCKER_WAIT = 5
 DOCKERFILE_DATASTORES = [
-    "mssql",
     "postgres",
     "mysql",
     "mongodb",
     "mariadb",
+    "mssql",
 ]
 EXTERNAL_DATASTORE_CONFIG = {
     "snowflake": ["SNOWFLAKE_TEST_URI"],
@@ -114,7 +114,9 @@ def seed_initial_data(
     _run_cmd_or_err('echo "Seeding initial data for all datastores..."')
     for datastore in datastores:
         if datastore in DOCKERFILE_DATASTORES:
-            setup_path = f"tests/ops/integration_tests/setup_scripts/{datastore}_setup.py"
+            setup_path = (
+                f"tests/ops/integration_tests/setup_scripts/{datastore}_setup.py"
+            )
             _run_cmd_or_err(
                 f'echo "Attempting to create schema and seed initial data for {datastore} from {setup_path}..."'
             )